+++ HISTORY.md
@@ -4,6 +4,7 @@ History of features added to LiChess Tools in time (reversed order)
 
 - support for multiple glyphs on the same move
 - disable Live Annotate option
+- Chess.com sound theme
 
 02 Jun